#96 move GraphicalDsl to new package
