+++ api/persistence/src/test/scala/de/htwg/zeta/persistence/behavior/RepositoryBehavior.scala
@@ -7,13 +7,13 @@ import de.htwg.zeta.common.models.entity.Filter
 import de.htwg.zeta.common.models.entity.FilterImage
 import de.htwg.zeta.common.models.entity.Generator
 import de.htwg.zeta.common.models.entity.GeneratorImage
-import de.htwg.zeta.common.models.entity.GraphicalDsl
 import de.htwg.zeta.common.models.entity.GraphicalDslRelease
 import de.htwg.zeta.common.models.entity.Log
 import de.htwg.zeta.common.models.entity.Settings
 import de.htwg.zeta.common.models.entity.TimedTask
 import de.htwg.zeta.common.models.entity.User
 import de.htwg.zeta.common.models.modelDefinitions.model.GraphicalDslInstance
+import de.htwg.zeta.common.models.project.GraphicalDsl
 import de.htwg.zeta.persistence.fixtures.AccessAuthorisationFixtures
 import de.htwg.zeta.persistence.fixtures.BondedTaskFixtures
 import de.htwg.zeta.persistence.fixtures.EventDrivenTaskFixtures
