+++ api/server/app/de/htwg/zeta/server/generator/model/diagram/Diagram.scala
@@ -1,6 +1,6 @@
 package de.htwg.zeta.server.generator.model.diagram
 
-import de.htwg.zeta.common.models.entity.GraphicalDsl
+import de.htwg.zeta.common.models.project.GraphicalDsl
 import de.htwg.zeta.server.generator.model.diagram.action.ActionGroup
 import de.htwg.zeta.server.generator.model.diagram.edge.Edge
 import de.htwg.zeta.server.generator.model.diagram.node.Node
