+++ api/server/app/de/htwg/zeta/server/generator/parser/SprayParser.scala
@@ -29,9 +29,9 @@ import de.htwg.zeta.server.generator.model.shapecontainer.shape.geometrics.Text
 import de.htwg.zeta.server.generator.model.style.Style
 import de.htwg.zeta.server.generator.parser
 import grizzled.slf4j.Logging
-import de.htwg.zeta.common.models.entity.GraphicalDsl
 import de.htwg.zeta.common.models.modelDefinitions.concept.elements.MClass
 import de.htwg.zeta.common.models.modelDefinitions.concept.elements.MReference
+import de.htwg.zeta.common.models.project.GraphicalDsl
 
 /**
  * SprayParser
