+++ api/server/app/views/metamodel/MetaModelGraphicalEditor.scala.html
@@ -1,7 +1,7 @@
 @import java.util.UUID
 
-@import de.htwg.zeta.common.models.entity.GraphicalDsl
 @import de.htwg.zeta.common.models.entity.User
+@import de.htwg.zeta.common.models.project.GraphicalDsl
 @import play.api.libs.json.Json
