+++ api/server/app/views/webpage/WebpageDiagramsOverview.scala.html
@@ -1,7 +1,7 @@
-@import de.htwg.zeta.common.models.entity.GraphicalDsl
 @import de.htwg.zeta.common.models.entity.User
 @import de.htwg.zeta.common.models.modelDefinitions.concept.MetaModelShortInfo
 @import de.htwg.zeta.common.models.modelDefinitions.model.GraphicalDslInstanceShortInfo
+@import de.htwg.zeta.common.models.project.GraphicalDsl
 
 @(user: Option[User], metaModels: Seq[MetaModelShortInfo], metaModel: Option[GraphicalDsl], modelInstances: Seq[GraphicalDslInstanceShortInfo])
